chore: getParsedResourcesMap fixes
